Structure project files
