+++ pydebloatx/icon.ico
