+++ pydebloatx/img_res.py
